Split DIGI applet into CAT, DAX, TCI, IQ (#1627)
The DIGI applet bundled four unrelated services into one tile.  Each now
gets its own toggle button, drag-reorder slot, and float/dock affordance,
so users can show only what they need and arrange them independently.
TCI gains a per-channel RX gain backed by TciRxGain1-4 settings, keeping
DAX and TCI audio chains fully independent.  Port-in-use failures are
now surfaced honestly instead of leaving the Enable button falsely green.
