+++ CMakeLists.txt
@@ -503,7 +503,10 @@ set(GUI_SOURCES
     src/gui/PhoneCwApplet.cpp
     src/gui/PhoneApplet.cpp
     src/gui/EqApplet.cpp
-    src/gui/CatApplet.cpp
+    src/gui/CatControlApplet.cpp
+    src/gui/DaxApplet.cpp
+    src/gui/TciApplet.cpp
+    src/gui/DaxIqApplet.cpp
     src/gui/MqttApplet.cpp
     src/gui/MeterSlider.cpp
     src/gui/PhaseKnob.cpp
