+++ src/gui/MainWindow.cpp
@@ -24,7 +24,10 @@
 #include "PhoneCwApplet.h"
 #include "PhoneApplet.h"
 #include "EqApplet.h"
-#include "CatApplet.h"
+#include "CatControlApplet.h"
+#include "DaxApplet.h"
+#include "TciApplet.h"
+#include "DaxIqApplet.h"
 #include "AntennaGeniusApplet.h"
 #include "RadioSetupDialog.h"
 #include "NetworkDiagnosticsDialog.h"
@@ -1992,14 +1995,28 @@ MainWindow::MainWindow(QWidget* parent)
                 QString("/tmp/AetherSDR-CAT-%1").arg(kLetters[i]));
         }
     }
-    m_appletPanel->catApplet()->setRadioModel(&m_radioModel);
-    m_appletPanel->catApplet()->setRigctlServers(m_rigctlServers, kCatChannels);
-    m_appletPanel->catApplet()->setRigctlPtys(m_rigctlPtys, kCatChannels);
-    m_appletPanel->catApplet()->setAudioEngine(m_audio);
+    m_appletPanel->catControlApplet()->setRadioModel(&m_radioModel);
+    m_appletPanel->catControlApplet()->setRigctlServers(m_rigctlServers, kCatChannels);
+    m_appletPanel->catControlApplet()->setRigctlPtys(m_rigctlPtys, kCatChannels);
+    m_appletPanel->daxApplet()->setRadioModel(&m_radioModel);
+    m_appletPanel->daxIqApplet()->setRadioModel(&m_radioModel);
 #ifdef HAVE_WEBSOCKETS
     m_tciServer = new TciServer(&m_radioModel, this);
     m_tciServer->setAudioEngine(m_audio);
-    m_appletPanel->catApplet()->setTciServer(m_tciServer);
+    m_appletPanel->tciApplet()->setRadioModel(&m_radioModel);
+    m_appletPanel->tciApplet()->setTciServer(m_tciServer);
+
+    // TCI applet sliders → TciServer gain setters
+    connect(m_appletPanel->tciApplet(), &TciApplet::tciRxGainChanged,
+            m_tciServer, &TciServer::setRxChannelGain);
+    connect(m_appletPanel->tciApplet(), &TciApplet::tciTxGainChanged,
+            m_tciServer, &TciServer::setTxGain);
+
+    // TciServer level signals → TCI applet meters
+    connect(m_tciServer, &TciServer::rxLevel,
+            m_appletPanel->tciApplet(), &TciApplet::setTciRxLevel);
+    connect(m_tciServer, &TciServer::txLevel,
+            m_appletPanel->tciApplet(), &TciApplet::setTciTxLevel);
 
     // Wire slice state changes → TCI broadcasts
     connect(&m_radioModel, &RadioModel::sliceAdded, this, [this](SliceModel* s) {
@@ -2028,12 +2045,12 @@ MainWindow::MainWindow(QWidget* parent)
 #endif
 
 #if defined(Q_OS_MAC) || defined(HAVE_PIPEWIRE)
-    // DAX enable button in CatApplet → start/stop DAX bridge
-    connect(m_appletPanel->catApplet(), &CatApplet::daxToggled,
+    // DAX enable button in DaxApplet → start/stop DAX bridge
+    connect(m_appletPanel->daxApplet(), &DaxApplet::daxToggled,
             this, [this](bool on) {
         if (on) {
-            if (!startDax() && m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setDaxEnabled(false);
+            if (!startDax() && m_appletPanel && m_appletPanel->daxApplet())
+                m_appletPanel->daxApplet()->setDaxEnabled(false);
         } else {
             stopDax();
         }
@@ -3178,8 +3195,8 @@ void MainWindow::buildMenuBar()
                 else if (!on && m_rigctlServers[i] && m_rigctlServers[i]->isRunning())
                     m_rigctlServers[i]->stop();
             }
-            if (m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setTcpEnabled(on);
+            if (m_appletPanel && m_appletPanel->catControlApplet())
+                m_appletPanel->catControlApplet()->setTcpEnabled(on);
         }
     });
 
@@ -3210,8 +3227,8 @@ void MainWindow::buildMenuBar()
                 else if (!on && m_rigctlPtys[i] && m_rigctlPtys[i]->isRunning())
                     m_rigctlPtys[i]->stop();
             }
-            if (m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setPtyEnabled(on);
+            if (m_appletPanel && m_appletPanel->catControlApplet())
+                m_appletPanel->catControlApplet()->setPtyEnabled(on);
         }
     });
 #endif
@@ -3232,8 +3249,8 @@ void MainWindow::buildMenuBar()
             } else if (!on && m_tciServer->isRunning()) {
                 m_tciServer->stop();
             }
-            if (m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setTciEnabled(on);
+            if (m_appletPanel && m_appletPanel->tciApplet())
+                m_appletPanel->tciApplet()->setTciEnabled(on);
         }
 #endif
     });
@@ -3259,12 +3276,12 @@ void MainWindow::buildMenuBar()
         s.save();
         if (m_radioModel.isConnected()) {
             if (on) {
-                if (startDax() && m_appletPanel && m_appletPanel->catApplet())
-                    m_appletPanel->catApplet()->setDaxEnabled(true);
+                if (startDax() && m_appletPanel && m_appletPanel->daxApplet())
+                    m_appletPanel->daxApplet()->setDaxEnabled(true);
             } else {
                 stopDax();
-                if (m_appletPanel && m_appletPanel->catApplet())
-                    m_appletPanel->catApplet()->setDaxEnabled(false);
+                if (m_appletPanel && m_appletPanel->daxApplet())
+                    m_appletPanel->daxApplet()->setDaxEnabled(false);
             }
         }
     });
@@ -4430,8 +4447,8 @@ void MainWindow::onConnectionStateChanged(bool connected)
                              << "on port" << (basePort + i);
                 }
             }
-            if (m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setTcpEnabled(true);
+            if (m_appletPanel && m_appletPanel->catControlApplet())
+                m_appletPanel->catControlApplet()->setTcpEnabled(true);
         }
         // Auto-start 8-channel CAT virtual serial ports if enabled
         if (as.value("AutoStartCAT", "False").toString() == "True") {
@@ -4442,8 +4459,8 @@ void MainWindow::onConnectionStateChanged(bool connected)
                              << m_rigctlPtys[i]->symlinkPath();
                 }
             }
-            if (m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setPtyEnabled(true);
+            if (m_appletPanel && m_appletPanel->catControlApplet())
+                m_appletPanel->catControlApplet()->setPtyEnabled(true);
         }
 #ifdef HAVE_WEBSOCKETS
         // Auto-start TCI WebSocket server if enabled
@@ -4451,10 +4468,14 @@ void MainWindow::onConnectionStateChanged(bool connected)
             if (m_tciServer && !m_tciServer->isRunning()) {
                 int tciPort = as.value("TciPort", "50001").toInt();
                 m_tciServer->start(static_cast<quint16>(tciPort));
-                qDebug() << "AutoStart: TCI on port" << tciPort;
+                qDebug() << "AutoStart: TCI on port" << tciPort
+                         << " running=" << m_tciServer->isRunning();
             }
-            if (m_appletPanel && m_appletPanel->catApplet())
-                m_appletPanel->catApplet()->setTciEnabled(true);
+            // Only light up the Enable button if the server actually bound —
+            // otherwise the UI shows a green "Enable" while the port is in use.
+            if (m_appletPanel && m_appletPanel->tciApplet())
+                m_appletPanel->tciApplet()->setTciEnabled(
+                    m_tciServer && m_tciServer->isRunning());
         }
 #endif
         // Populate XVTR bands after radio status settles, and refresh
@@ -4484,8 +4505,8 @@ void MainWindow::onConnectionStateChanged(bool connected)
         // be registered in PanadapterStream yet.
         if (AppSettings::instance().value("AutoStartDAX", "False").toString() == "True") {
             QTimer::singleShot(3000, this, [this]() {
-                if (startDax() && m_appletPanel && m_appletPanel->catApplet())
-                    m_appletPanel->catApplet()->setDaxEnabled(true);
+                if (startDax() && m_appletPanel && m_appletPanel->daxApplet())
+                    m_appletPanel->daxApplet()->setDaxEnabled(true);
             });
         }
 #endif
@@ -7877,28 +7898,28 @@ bool MainWindow::startDax()
     connect(m_radioModel.panStream(), &PanadapterStream::iqDataReady,
             &m_radioModel.daxIqModel(), &DaxIqModel::feedRawIqPacket);
 
-    // Wire DAX IQ level meters to DIGI applet
+    // Wire DAX IQ level meters to DAX IQ applet
     connect(&m_radioModel.daxIqModel(), &DaxIqModel::iqLevelReady,
-            m_appletPanel->catApplet(), &CatApplet::setDaxIqLevel);
+            m_appletPanel->daxIqApplet(), &DaxIqApplet::setDaxIqLevel);
 
-    // Wire DAX IQ enable/disable/rate from DIGI applet to DaxIqModel
-    connect(m_appletPanel->catApplet(), &CatApplet::iqEnableRequested,
+    // Wire DAX IQ enable/disable/rate from DAX IQ applet to DaxIqModel
+    connect(m_appletPanel->daxIqApplet(), &DaxIqApplet::iqEnableRequested,
             &m_radioModel.daxIqModel(), &DaxIqModel::createStream);
-    connect(m_appletPanel->catApplet(), &CatApplet::iqDisableRequested,
+    connect(m_appletPanel->daxIqApplet(), &DaxIqApplet::iqDisableRequested,
             &m_radioModel.daxIqModel(), &DaxIqModel::removeStream);
-    connect(m_appletPanel->catApplet(), &CatApplet::iqRateChanged,
+    connect(m_appletPanel->daxIqApplet(), &DaxIqApplet::iqRateChanged,
             &m_radioModel.daxIqModel(), &DaxIqModel::setSampleRate);
 
     // Wire DAX level meters
     connect(m_daxBridge, &DaxBridge::daxRxLevel,
-            m_appletPanel->catApplet(), &CatApplet::setDaxRxLevel);
+            m_appletPanel->daxApplet(), &DaxApplet::setDaxRxLevel);
     connect(m_daxBridge, &DaxBridge::daxTxLevel,
-            m_appletPanel->catApplet(), &CatApplet::setDaxTxLevel);
+            m_appletPanel->daxApplet(), &DaxApplet::setDaxTxLevel);
 
     // Wire DAX gain sliders
-    connect(m_appletPanel->catApplet(), &CatApplet::daxRxGainChanged,
+    connect(m_appletPanel->daxApplet(), &DaxApplet::daxRxGainChanged,
             m_daxBridge, &DaxBridge::setChannelGain);
-    connect(m_appletPanel->catApplet(), &CatApplet::daxTxGainChanged,
+    connect(m_appletPanel->daxApplet(), &DaxApplet::daxTxGainChanged,
             m_daxBridge, &DaxBridge::setTxGain);
 
     // Apply saved gains to the bridge
